LayoutTests/fast/canvas/canvas-getImageData-negative-source-expected.txt

Test the handling of negative sourceRect offset in getImageData().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[0] is 0
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 0
PASS imgdata[60] is 0
PASS imgdata[61] is 128
PASS imgdata[62] is 0
PASS imgdata[63] is 255
PASS imgdata[64] is 255
PASS imgdata[65] is 0
PASS imgdata[66] is 0
PASS imgdata[67] is 255
PASS successfullyParsed is true

TEST COMPLETE
